Front End Forms
Created front end forms for the different user stories as well as the forms for approving and denying them.

For managers,the data attached to aform will be loaded into the specific input fields. Managers should not be able to modify forms, only clicking approve or deny after viewing the submitted forms.