Worked on Login Component
Able to load the login page. The login and register forms are able to
switch between one another.

Unsure if we are able to send data yet. The buttons are attempting to
post data but since we are not sending any HTTP Requests, i receive an
error so at least that is working out.